Slightly changed galaxy background
Changed the galaxy background to match the .ai file in the team Google Drive folder. The differences are very slight, but it's more consistent this way. Also removed the .svg from the img folder: it didn't need to be there and a vector graphic can be found in the Google Drive folder. Only rasterized images should be used for the galaxy background because the vector graphic has lots of paths.